+++ .gitignore
@@ -40,6 +40,6 @@ Scripts/gcovr
 Scripts/jenkins.py
 Scripts/ocunit2junit
 __pycache__/
-awsconfiguration.json
-amplifyconfiguration.json
+*awsconfiguration.json
+*amplifyconfiguration.json
 credentials-mc.json
